server/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3
yt-dlp==2024.12.3
gunicorn==23.0.0
